Refactor duplicate code to a common static class. Also move the workspaceInfo parser to a subpackage.
